+++ src/client/src/openfinEpics.ts
@@ -1,28 +1,12 @@
-import { combineEpics, ofType, StateObservable } from 'redux-observable'
+import { combineEpics, ofType } from 'redux-observable'
 import { bindCallback } from 'rxjs'
 import { filter, map, mergeMap, switchMapTo, takeUntil, tap, withLatestFrom } from 'rxjs/operators'
 import { ApplicationEpic } from './ApplicationEpic'
-import { GlobalState } from './combineReducers'
 import { CONNECT_SERVICES, DISCONNECT_SERVICES } from './connectionActions'
-import { CurrencyPairReducerState } from './currencyPairsOperations'
-import { OpenFin } from './services'
-import { Direction, SpotPriceTick, Trades } from './types'
+import { connectBlotterServiceToOpenFinEpic } from './services/openFin'
+import { Direction, SpotPriceTick } from './types'
 import { CurrencyPair } from './types/currencyPair'
-import { ACTION_TYPES as BLOTTER_ACTION_TYPES } from './ui/blotter/actions'
 import { SpotTileActions } from './ui/spotTile/actions'
-
-const subscribeOpenFinToBlotterData = (openFin: OpenFin, state$: StateObservable<GlobalState>) => () => {
-  const trades: Trades = state$.value.blotterService.trades
-  const currencyPairs: CurrencyPairReducerState = state$.value.currencyPairs
-  const cb = (msg: any, uuid: string) => openFin.sendAllBlotterData(uuid, trades, currencyPairs)
-  openFin.addSubscription('fetch-blotter', cb)
-}
-
-export const connectBlotterServiceToOpenFinEpic: ApplicationEpic = (action$, state$, { openFin }) =>
-  action$.pipe(
-    ofType(BLOTTER_ACTION_TYPES.BLOTTER_SERVICE_NEW_TRADES),
-    tap(subscribeOpenFinToBlotterData(openFin, state$))
-  )
 
 const addRatePrecisionToPrice = (currencyData: Map<string, CurrencyPair>, price: SpotPriceTick) => {
   return {
