+++ src/client/src/services/openFin/epics.ts
@@ -1,0 +1,21 @@
+import { ofType, StateObservable } from 'redux-observable'
+import { tap } from 'rxjs/operators'
+import { ApplicationEpic } from '../../ApplicationEpic'
+import { GlobalState } from '../../combineReducers'
+import { CurrencyPairReducerState } from '../../currencyPairsOperations'
+import { Trades } from '../../types'
+import { ACTION_TYPES as BLOTTER_ACTION_TYPES } from '../../ui/blotter'
+import OpenFin from './openFin'
+
+const subscribeOpenFinToBlotterData = (openFin: OpenFin, state$: StateObservable<GlobalState>) => () => {
+  const trades: Trades = state$.value.blotterService.trades
+  const currencyPairs: CurrencyPairReducerState = state$.value.currencyPairs
+  const cb = (msg: any, uuid: string) => openFin.sendAllBlotterData(uuid, trades, currencyPairs)
+  openFin.addSubscription('fetch-blotter', cb)
+}
+
+export const connectBlotterServiceToOpenFinEpic: ApplicationEpic = (action$, state$, { openFin }) =>
+  action$.pipe(
+    ofType(BLOTTER_ACTION_TYPES.BLOTTER_SERVICE_NEW_TRADES),
+    tap(subscribeOpenFinToBlotterData(openFin, state$))
+  )
