+++ src/client/src/services/openFin/index.ts
@@ -1,1 +1,2 @@
 export { default as OpenFin } from './openFin'
+export { connectBlotterServiceToOpenFinEpic } from './epics'
