+++ src/client/src/ui/blotter/index.ts
@@ -2,3 +2,4 @@ export { default as Blotter } from './Blotter'
 export { default as BlotterContainer } from './BlotterContainer'
 export { default as blotterReducer } from './reducer'
 export { default as blotterEpic } from './epics'
+export { ACTION_TYPES, BlotterActions } from './actions'
